requirements.txt
google-auth==1.35.0
google-api-python-client==2.17.0
firebase-admin==5.0.2
python-dotenv==0.19.0
python-docx==0.8.11
